+++ pkg/podgrouper/podgroup/handler.go
@@ -117,6 +117,7 @@ func (h *Handler) createPodGroupForMetadata(podGroupMetadata Metadata) *scheduli
 			schedulingv2alpha2.SubGroup{
 				Name:      subGroup.Name,
 				MinMember: subGroup.MinAvailable,
+				Parent:    subGroup.Parent,
 			})
 	}