+++ pkg/podgrouper/podgroup/metadata.go
@@ -11,6 +11,7 @@ import (
 type SubGroupMetadata struct {
 	Name           string
 	MinAvailable   int32
+	Parent         *string
 	PodsReferences []*types.NamespacedName
 }